BBQ debug info dump

+++ VT_MSPM0G3507_A/src/Middleware/BQ769x2/BQ76952.cpp
@@ -36,14 +36,14 @@ bool BQ76952::setRegister(BQ769X2_PROTOCOL::RegAddr reg_addr, uint32_t reg_data,
     return BQ769X2_PROTOCOL::setRegister(*i2c_controller, i2c_addr, reg_addr, reg_data, datalen);
 }
 
-bool BQ76952::I2C_ReadReg(uint8_t reg_addr, void *reg_data, uint8_t count) {
+bool BQ76952::I2C_ReadReg(uint16_t reg_addr, void *reg_data, uint8_t count) {
     return BQ769X2_PROTOCOL::I2C_ReadReg(*i2c_controller, i2c_addr, reg_addr, (uint8_t*)reg_data, count);
 }
 
-bool BQ76952::I2C_WriteReg(uint8_t reg_addr, void *reg_data, uint8_t count){
+bool BQ76952::I2C_WriteReg(uint16_t reg_addr, void *reg_data, uint8_t count){
     return BQ769X2_PROTOCOL::I2C_WriteReg(*i2c_controller, i2c_addr, reg_addr, (uint8_t*)reg_data, count);
 }
 
-bool BQ76952::I2C_WriteReg(uint8_t reg_addr, uint8_t data){
+bool BQ76952::I2C_WriteReg(uint16_t reg_addr, uint8_t data){
     return BQ769X2_PROTOCOL::I2C_WriteReg(*i2c_controller, i2c_addr, reg_addr, &data, 1);
 }
