+++ VT_MSPM0G3507_A/src/Middleware/BQ769x2/BQ76952.hpp
@@ -52,9 +52,9 @@ public:
 
     bool setRegister(BQ769X2_PROTOCOL::RegAddr reg_addr, uint32_t reg_data, uint8_t datalen);
 
-    bool I2C_ReadReg(uint8_t reg_addr, void *reg_data, uint8_t count);
-    bool I2C_WriteReg(uint8_t reg_addr, void *reg_data, uint8_t count);
-    bool I2C_WriteReg(uint8_t reg_addr, uint8_t data);
+    bool I2C_ReadReg(uint16_t reg_addr, void *reg_data, uint8_t count);
+    bool I2C_WriteReg(uint16_t reg_addr, void *reg_data, uint8_t count);
+    bool I2C_WriteReg(uint16_t reg_addr, uint8_t data);
 };
